Rearranged code to make new classes exportable.
